HYD-2405: ManagedTarget.inode_count should be BigIntegerField
With large (> 3.6TB) MDTs, the default bytes_per_inode value would
yield an inode_count that overflows IntegerField in postgresql.

Also updates the cluster simulator to return a more realistic
inode_count after formatting targets. Having this in place validates
the db change as we can persist a fake 32TB MDT with the fix.

+++ chroma-manager/chroma_core/models/target.py
@@ -76,7 +76,7 @@ class ManagedTarget(StatefulObject):
     inode_size = models.IntegerField(null = True, blank = True, help_text = "Size in bytes per inode")
     bytes_per_inode = models.IntegerField(null = True, blank = True, help_text = "Constant used during formatting to "
                                           "determine inode count by dividing the volume size by ``bytes_per_inode``")
-    inode_count = models.IntegerField(null = True, blank = True, help_text = "The number of inodes in this target's"
+    inode_count = models.BigIntegerField(null = True, blank = True, help_text = "The number of inodes in this target's"
                                       "backing store")
 
     def get_hosts(self, primary=True):
